sparvi-core/requirements.txt
sqlalchemy>=1.4.0
pandas>=1.3.0
numpy>=1.20.0
click>=8.0.0
pyyaml>=6.0
rich>=10.0.0
duckdb>=1.0.0